requirements.txt
app==0.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
repoze.lru==0.7
Routes==2.5.1
six==1.16.0
Werkzeug==3.0.3
